+++ src/adapters/SequelizeAdapter.js
@@ -1,8 +1,0 @@
-import DefaultAdapter from './DefaultAdapter.js';
-
-/* eslint-disable no-unused-vars */
-export default class SequelizeAdapter extends DefaultAdapter {
-  build(Model, props) {
-    return Model.build(props);
-  }
-}